Remove dependency on the versions plugin
